Woo Product Create: Use own fetch instead of http
This commit changes to use the request module based on wpcom.request,
but it uses it still in the data layer rather than in an action thunk.

This is mostly because the changes needed for the data layer http module
to make it work in extensions is still being looked at and determined.